Todo finalizado falta comentar codigo y readme

+++ view/usuarios.php
@@ -86,14 +86,18 @@ if ($_SESSION['email']=="") {
                 echo  "<th class='blue'>Tipo</th>";
                 echo  "</tr>";
                 foreach ($comprobacion as $row) {
-                    echo "<tr>";
-                    echo "<td class='gris'>{$row['email']}</td>";
-                    echo "<td class='gris'>{$row['nombre']}</td>";
-                    echo "<td class='gris'>{$row['apellido']}</td>";
-                    echo "<td class='gris'>{$row['tipo']}</td>";
-                    echo "<td><button type='submit'><a type='button' href='../view/formmodificarusuario.php?email={$row['email']}'>Modificar usuario</a></button></td>";
-                    echo "<td><button class='buttononline' type='submit'><a type='button' href='../proceses/eliminarusuario.php?email={$row['email']}'>Eliminar usuario</a></button></td>";
-                    echo "<tr>";
+                    if ($row["nombre"]==$_SESSION['nombre']) {
+                        # code...
+                    }else{
+                        echo "<tr>";
+                        echo "<td class='gris'>{$row['email']}</td>";
+                        echo "<td class='gris'>{$row['nombre']}</td>";
+                        echo "<td class='gris'>{$row['apellido']}</td>";
+                        echo "<td class='gris'>{$row['tipo']}</td>";
+                        echo "<td><button type='submit'><a type='button' href='../view/formmodificarusuario.php?email={$row['email']}'>Modificar usuario</a></button></td>";
+                        echo "<td><button class='buttononline' type='submit'><a type='button' href='../proceses/eliminarusuario.php?email={$row['email']}'>Eliminar usuario</a></button></td>";
+                        echo "<tr>";
+                    }
                 }
                 echo  "</div>";
                 echo  "</div>";
